pretty code

+++ lab2/mul_and_div.c
@@ -9,17 +9,17 @@
 
 int main()
 {
-    srand(time(NULL));
+    srand((unsigned int) time(NULL));
 
     uint256_t dig1 = {0}, dig2 = {0};
     uint256_t res2 = {0}, res3 = {0};
     uint512_t res1 = {0};
+
     Generate(dig1);
     Generate(dig2);
+
     Mul(dig1, dig2, res1);
     Div(res1, dig2, res2, res3);
-
-
 
     Print(dig1);
     Print(dig2);
@@ -31,14 +31,12 @@ int main()
     }
     printf("\n");
 
-
     Print(res2);
     Print(res3);
 
     printf("result of compare: %d\n", Cmp(res2, dig1));
     uint256_t zero = {0};
     printf("result of compare: %d\n", Cmp(res3, zero));
-
 
     return 0;
 }
